Layout refinements
I noticed that when not logged in the new layout forthe "fitness leaderborad" would stay aligned left so I modified the autorized/not code to show 2 different layouts to keep that centered if user is not logged in. Alos updated some classes to center the user avatar o the profile page and some margins and paddings.